Contribution from Michael Peters (wonko) from CPAN RT.
Add support for specificity calculation. I looked up a number of different implementations for this that were available on other libraries - this seemed to be equally if not more complete, and shorter.

Including his test verbatim, including his patch but altering it slightly to comply with existing conventions within the lib.

Original contribution here:
https://rt.cpan.org/Public/Bug/Display.html?id=59064



git-svn-id: http://repo/sysadmin/perlmodules/CSS-Inliner@2768 49aadcc1-c794-da11-82e1-00902723a867