Contribution from Michael Peters (wonko) from CPAN RT.
The same thread from CPAN that included the specificity patch also included support for applying the cascade, along with a test.

I've merged that code into the current form of the tree. There have been many changes so the patch could not be applied directly.

I've removed references to refaddr and instead used the address method for HTML::Element which basically provides the same type of idea of uniquely identifying an element, but via a scalar.

We probably want to refactor this a little bit, but I've added the tests that came along with the patch and it works as described.

I've additionally added an acid test from HTML::Query to test out the inlinification. At this point I realized we have a broken grandfather operator in that library as well as a broken universal selector.

Additionally there is a precendence order bug within the included patch as revealed by the acid test. We'll work on fixing those up text.


git-svn-id: http://repo/sysadmin/perlmodules/CSS-Inliner@2770 49aadcc1-c794-da11-82e1-00902723a867