MM-4027
first round of updates to avoid croaking while handlings various selectors and selector values.

More work needed, currently Simple.pm prints its errors as warnings, which is less than useful right now. It will need to be updated to aggregate those as well as provide a method for retrieving them - that or 
perhaps the call will also be wrapped in an eval in order to trap that particular value for later.

We will also need to provide 2 different types of error reports. Those related to actual errors vs simple content warnings, just not sure what fits in which category and how...



git-svn-id: http://repo/sysadmin/perlmodules/CSS-Inliner@3200 49aadcc1-c794-da11-82e1-00902723a867