MM-4571
addition of filter and unfilter content methods for fetched content

Currently does not work right - I do not know why.

The attached script acts as a passthru and demonstrates the feature, but it's somehow altering the content in some fashion - I cannot find out how.

When I put print statements in the filter (which is actually really useful) I see exactly what I expect - so not sure what's up yet.

In any case the code that is to be passed as a filter needs to be determined. Is that what is in MessageTweak? Or are we going to need copy/paste of more methods in the project?




git-svn-id: http://repo/sysadmin/perlmodules/CSS-Inliner@3269 49aadcc1-c794-da11-82e1-00902723a867