MM-4520
Update inliner to accept a parameter for the character set of a file reference intended to be inlined.

Update inliner to dynamically determine a character set for a remote document on a webserver.

If a character set is not provided we just behave as we did before - basically hope for the best...

Update the comments within "read" to inform the programmer that they are responsible for making sure they
properly identify the character set within perl before calling "read" as it is too far down the chain to be
concerned with this problem.



git-svn-id: http://repo/sysadmin/perlmodules/CSS-Inliner@3671 49aadcc1-c794-da11-82e1-00902723a867