MM-8488
Update module description and copyright with proper year. This is a maintained module, so keeping these up to date
is something we need to spend more time on...

Additionally strip out "post_fetch_filter". This was an undocumented feature that was used internally by the module
sponsor. It is no longer used there and may now be removed.