MM-8867
While performing some inlining of some various websites we discovered that HTML5 tags are entirely unsupported
by the underlying perl HTML Parser. To resolve this problem a sprint was done whereby a type of "relaxed" parsing
is enabled within the underlying Parser which disables specification conformance within any parsed documents.

This not only will allow HTML5 documents to be handled (although not in a specification conformant way) but also
will allow "in the wild" type documents to be handled as well.

To add the above feature HTML::TreeBuilder had to be overridden locally within the project. Any handles to
HTML::TreeBuilder are upgraded to our own internal TreeBuilder.

Additionally while here it was discovered that indentation was broken for all resulting HTML. This was corrected
for all HTML except those handled by relaxed mode. Relaxed parsing may or may not return properly indented html,
depending on whether or not it's entirely built of valid tags or not.

Finally a number of tests were updated, and the MANIFEST file was updated to be accurate again.