MM-8867
Additional refinements to relaxed parsing mode

HTML::TreeBuilder at it's core is a container. If we want to only print the interior contents
(which in this case we do, because the container is known to be spurious) the only real option
is to hack the HTML::TreeBuilder instance and indicate that the element is an empty literal.

By doing so we suppress both a start and end tag, but are left with some goofy indenting which
we later can take care of manually when we produce the final "inlined" text.

This approach is hacky, but I can find no other possible way through which to implement the
necessary functionality to pull this off. This seems to be the only possible way through which
HTML5 and sloppy HTML documents can be parsed.

Noticed that acidtest.t was not actually doing anything, so I removed the secondary test there,
as there is a dedicated test for determining if relaxed mode is working.

Renamed crazytags to relaxed, as the feature now has a name.