MM-8867
Additional refinements to relaxed mode, which now very tightly coupled with the
implicit_tags setting for TreeBuilder.

It's pretty difficult to handle all the permutations of these flags while also
handling an argument containing a treebuilder reference - it's not clear which
argument would actually take precedence. Rather than deal with that problem
I am simply disallowing it for now, so the use of the html_tree and relaxed
flags will now throw an error if used together.