MM-8867
The regex that was used to protect "declarations" was nowhere near specific enough -
turned out it was matching the IE if/endif declarations too - which were being handled
correctly previously, but now were broken.

Investigated further and it appears that there was special considering for doctype
within HTML::TreeBuilder, so we only specifically look to undo that particular work, as
no other tags seemingly got this kind of special treatment.