MM-8867
The previous technique of removing the spurious wrapping HTML tags within
the relaxed content output failed - in the event that you had trailing content
outside the html area it would strip that content too, and not comply with the
new "leave as much alone as possible" philosophy I am trying to implement.

Reverted back to converting the wrapping HTML block to a wrapping literal element
which is a little hacky, but a lot closer to generating a "legit" result. There
is still some stripping of whitespace that has to be done, but it's relatively
straightforward.

Given the way that TreeBuilder is implemented there is a risk here that the
literal tag could be made more strict, and thus break what I have here, but I think
that risk is small given how stable this specific code has been within TreeBuilder.