MM-8867
Rework the POD documentation to use consistent formatting throughout

Remove line of code that was stripping out the wrapping html tags around
a document - I cannot recall why I added this, but it does not appear to
be serving a useful purpose

Add additional code comments

When retrieving a remote document ("fetching") use relaxed parsing
to perform the initial parse - it is far less destructive than the base
TreeBuilder which was used before

Relax the identification queries for finding style/link tags within
the document(s). Previously we required the standard "text/css" type
to be present for style tags, and we required the standard rel
"stylesheet" and type "text/css" for link tags. Updated such that finding
these tags is good enough to try to inline them.

Update the validator method such that the expected counts for wrapping
html nodes reflects the recent change to TreeBuilder, it's a bit more
intuitive now.