MM-12877
Rework the entities test, which previously tested whether or not default entities (<>&)
remained unescaped, along with testing whether the entities argument was used for
escaping the result.

The module no longer supports automatic entity encoding, even when handling potentially
problemantic characters (for example < and > outside of tag names). This test is now
updated to ensure that such characters remain untouched, and the reference to custom
entity encoding (which was done through the entities argument) has been removed.