MM-12877
Refine the acidtestutf8 test, swap out the last of the combining characters, as they are
much harder to work with.

Create the corresponding results file for the test, and finish the final logic for the comparison.

Finally finish up the logic for the test such that the inlined result is compared against the
stored result file.