MM-13248
Resolve issue with encoding extraction regex. Bumped version number to prepare for follow up release.

While here also noticed that a standalone HTML::TreeBuilder was instantiated within this method. This is now
replaced with a CSS::Inliner::TreeBuilder, and configured identically to the other instances within the class
for consistency purposes.