MM-XXXX
Fix embedded_style_block.t.

Related PR would always fail. Reason for this is that
LWP::Simple->get() does not pass a proper user agent, and is blocked
by a very large number of webservers, including rawgit.

The requirement, to fetch a remote file, is already accounted for
within this particular library. I basically removed the get() request,
and then shuffled some code around.

If Inliner is unable to request the file, we will now fail before
attempting to inline. Otherwise it will proceed as normal. We'll see
if there are other unstated issues that the previous approach was
dealing with.