Switch to array structure to retain the original order of css rules
Even for rules with different names, the order matters, e.g.

    border: none;
    border-bottom: 1px solid #ddd;

The border-bottom is "1px solid #ddd" and will be "none" if the order of
the 2 rules is reversed.

This commit changes data structure from unordered hash to array, to
retain the order of rules as much as it can.

Note that the internal order of "!important" rules is also retained,
although they are moved to the end of "normal" rules as a whole(quite
like before).